daemonization: terminate main process after init
Main process waits for signal (SIGUSR1) from the last forked child.
Timeout is set to 3 seconds, after which the main process exits with
non-zero (1) exit code.

Without this fix, we can encounter race condition with systemd when
starting the service. If the main process is terminated and
deamonization is not finished, systemd kills the whole process group.