tuned: new option --log, -l to select log file
If used without argument as --log or -l the tuned will log to file even if
running on foreground. It will not log to console. This is the default if
running as daemon.

If used with argument as --log FILE, -l FILE, the log will be stored to FILE.
This is also applicable for deamon.

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>