tuned: new option --pid, -P to write PID file
If used without argument as --pid or -P the tuned will write the PID file even
if running in foreground. The PID file is always written if running as daemon.

If used with argument as --pid FILE or -P FILE, the PID will be written to
FILE. This is also applicable for daemon.

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>