profiles: added SAP profile
Also added support for the sysctl tunings taken from the
environment and conditional tunings (i.e. not to change
the sysctl settings if the current value is greater than
the new value).

To use the environment variable SEM:
kernel.sem = $SEM

To set the kernel.shmall only if the current value is less
than the 1000000:
kernel.shmall = >$1000000

Both can be combined, e.g.:
kernel.shmall = >$SHMALL

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>