tuned: timing improvements
This improves responsiveness of Tuned in Python 2 interpreter with applied
patch for rhbz#917709. There is new global config parameter "sleep_interval"
which can control the polling. By default it is set to 1 second which is
still much better than 50 ms (on unpatched intepreter). Also the overhead
of the main thread was lowered if the dynamic tuning is disabled. The
dynamic tuning update interval is still controlled by the "update_interval"
global config parameter, but it has to be multiply of the "sleep_interval".

Resolves: rhbz#1028122

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>