balanced: used medium_power ALPM policy
This policy sets the link to the second lowest power
state (PARTIAL) when there is no I/O on the disk. This
mode is designed to allow transitions in link power states
(for example during times of intermittent heavy I/O and
idle I/O) with as small impact on performance as possible.

This may save significant amount of power in idle periods
of the link, especially on laptops. E.g. on x240 the average
AC power consumption dropped from approx. 5 W to only 2.5 W.

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>