tuned: Added support for >, < assignment modifiers in tuned.conf
These modifiers can specify when to assign new value considering
the old value.

E.g.:
readahead=>4096

means assign readahead 4096 only if the current value of readahaed
is lower than 4096 (i.e. the readahaed will be at least 4096).

Similarly for the < operator, it will assign the value only
if the current value is higher.

The modifiers are part of the value, not part of the equality
operator, thus the following is correct:
readahead= >4096

But the following isn't correct, but will probably also work
due to the nature how the value is parsed:
readahead=> 4096

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>