plugin_disk: readahead value can be now specified in sectors
E.g.:
[disk]
readahead=256 s

This will be internally transformed to 128 kB and has
same meaning as:

[disk]
readahead=128

Any suffix starting with "s" can be used for sectors,
e.g. "sects" or "sectors", etc. If there is no suffix
or the suffix doesn't start with "s", kB are used.
There must be at least one space between the number
and suffix.

Resolves: rhbz#1127127

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>