plugins: added scheduler plugin
Currently it is substitution for the rtctl tool,
but there are small differences in the syntax.

Format for the rules is:

group.GROUPNAME=RULE_PRIO:SCHED:PRIO:AFFINITY:REGEX

where RULE_PRIO defines tuned priority of the
rule. Rules are sorted according to priority.
This is needed for inheritence to be able
to reorder previously defined rule. For same
RULE_PRIO, rules should be processed in the order
there were defined (but this is Python interpreter
dependant). To disable inherited rule for GROUPNAME
use:

group.GROUPNAME=

SCHED must be one of:
 'f' for FIFO,
 'b' for batch,
 'r' for round robin,
 'o' for other,
 '*' means not to change.

AFFINITY is a hex number, see taskset(1) for details
about number of CPUs. The '*' means not to change.

REGEX is Python regex. It must match against output of

  ps -eo cmd

Any given process name may match more than one group.
Default priority and scheduling policy are taken from
the last matching rege.

You need to write 'group.' prefix before your
group name. This is because there may be added
other parameters to the plugin later.

Example of usage:

[scheduler]
group.kthreads=0:*:1:*:\[.*\]$
group.watchdog=0:f:99:*:\[watchdog.*\]

Resolves: rhbz#1100826

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>