profiles: reversed SYSCTL_POST="/etc/sysctl.d/* /etc/sysctl.conf"
This commit reversed it to read:
SYSCTL_POST="/etc/sysctl.conf /etc/sysctl.d/*"
The idea is to go from implicit to explicit, i.e. from defaults
to overrides in sysctl.d subdirectory. This behavior is
expected in RHEL-6.

But many distributions use the reverse approach (state before
this commit). The idea is that /etc/sysctl.d/* hosts default
settings provided by individual packages which can be overriden
by "master" /etc/sysctl.conf.

In Fedora (F20 checked) it is implemented by the following
symlink:
/etc/sysctl.d/99-sysctl.conf -> ../sysctl.conf

So it is also effectively reversed in comparison to RHEL-6.

As tuned-0.x is supported only in RHEL-6, we need to follow
RHEL-6 behaviour in tuned-1.0 branch. This change doesn't
affect tuned-2.x.

resolves: rhbz#1036049

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>