recommend: profile can be specified multiple times by using suffix
It needs to use unique ',ANYSTRING' suffix, everything after the last ','
will be stripped by the parser, e.g.:

[balanced,1]
/FILE1=RE1

[balanced,2]
/FILE2=RE2

This will set 'balanced' profile in case there is FILE1 matching RE1 or
FILE2 matching RE2 or both. There is no need to use the suffix
if the profile is specified only once.

Also improved description in the recommend.conf.

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>