ktune: do not revert sysctl settings on ktune stop
Now the sysctl settings are not reverted on ktune stop. This is to avoid cases
where the time window when the sysctl settings are reverted can cause unwanted
behaviour. E.g. it can temporary reset system limits on disk addition (because
ktune service is restarted by udev) which may cause some commands to fail.

It is still possible to switch to the old behaviour (reverting sysctl) by
exporting KTUNE_REVERT_SYSCTL and setting it to 1. This can be also
set per profile in the /etc/tune-profiles/PROFILE_NAME/ktune.sysconfig by
adding the following:

KTUNE_REVERT_SYSCTL=1

Resolves: rhbz#1111416

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>