plugin_disk: try to workaround embedded '/' in device names
It seems that some drivers return udev names with embedded slashes '/', e.g.:
sfdsk/gdssys34

When kernel creates sysfs nodes for them it replaces '/' by '!' not to cause
ambiguity in path, i.e. the device from previous example will become:
sfdsk!gdssys34

This commit add support for such cases. It checks whether there is embedded
'/' in the device name, if yes, it tries to construct sysfs path with '!'.
It checks whether the resulting sysfs path exists, if not, it tries the
version with embedded '/'. This is safety fallback not to cause regression
in some weird cases and it may be dropped in the future.

Related: rhbz#1353142

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>