plugin_bootloader: Rename _grub2_cfg_file attribute
Rename the _grub2_cfg_file attribute to _grub2_cfg_file_name. This is to
avoid the following pylint error message:
An attribute defined in tuned.plugins.plugin_bootloader line 30 hides this method.

The fact that the method was hidden didn't present itself in any way
during execution, because the method had been saved in a collection
(and accessed only from the collection after that) prior to becoming
hidden.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>