Introduce ${PROFILE_DIR} variable, drop workdir option
Introduced the ${PROFILE_DIR} constant, which can be used
in profiles and which expands to the directory in which the profile
configuration was found. It replaces the workdir option, which
was a bit problematic due to inheritance and the possibility
to change its value.

The ability to change the value of 'workdir' seems redundant,
as the same effect can be achieved by either specifying an absolute
path to files in unit configuration or by using a user defined
variable.

Note that now you need to specify an absolute path or use the PROFILE_DIR
constant in the initrd_add_dir and initrd_add_img options of the bootloader
plugin.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>