variables: fixed parsing of function parameters containing undefined variables
Fixed parsing of e.g.:
${f:a:${var1}:${var2}}

If var1 wasn't defined it was incorrectly parsed as:
f:a:${var1

I.e. it resulted in invalid number of parameters error, as the
called function was missing the second parameter.

Now with bot variables undefined it parses as:
f:a:${var1}:${var2}

I.e. it resulted in correct number of parameters even with
undefined variables.

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>