builtin functions: added virt_check function and support expansion in include
The virt_check builtin function expands to the first argument if
running inside virtual machine, otherwise it expands to the second
argument (when on bare metal or in case of error). It uses virt-what
as a backend.

So now the following can be used to lower fragmentation of Tuned profiles:

[main]
include=virtual-${f:virt_check:guest:host}

So it will include virtual-guest if running inside virtual machine,
otherwise it includes virtual-host.

Resolves: rhbz#1426654

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>