realtime: use skew_tick=1 in kernel cmdline
By default the timer tick will fire at the same time on all
CPUs. This causes the jiffies_lock to get contended in the
real-time kernel, as it's a raw spinlock. skew_tick=1 causes
the timer tick to fire at different times on all CPUs,
fixing the contention.

Note skew_tick=1 usage is going to be propagated to the
NFV profiles as well.

Resolves: rhbz#1447938

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>