Improve checking if we are rebooting or not
'systemctl is-system-running' is not enough to reliably check
if the system is shutting down. If the command is run when the system
is still starting up, it will report the state as 'starting', even if
the shutdown target is queued:
https://github.com/systemd/systemd/blob/5463fa0a88f95d2002858592578f9bf4e0d2660a/src/core/manager.c#L3457-L3465

On the other hand, 'systemctl list-jobs' should always show
shutdown.target or reboot.target etc., when the system is shutting
down, as long as DBus is still running.

Resolves: rhbz#1475571
Resolves: rhbz#1488369
Resolves: rhbz#1488517

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>