cpu: Support specifying alternative EPB values
Support specifying alternative Energy Performance Bias values.
The values are separated using the '|' character. For example,
if you have the following in your profile:

[cpu]
energy_perf_bias=powersave|power

then tuned will try to set EPB to 'powersave', and if that fails,
it will try to set it to 'power'.

Resolves: rhbz#1508468

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>