scheduler: Drop no_error argument in _set_affinity()
I believe it was originally added so that errors are not logged
when we fail to set the affinity of a short-lived process which
has already disappeared. This is no longer necessary, because we
always check if the process has disappeared, and log debug messages
instead of errors if it has.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>