scheduler: Use schedutils in _set_rt() and _get_rt()
Rewrite _set_rt() and _get_rt() to use schedutils instead of
command line tools.

Behaviour change: previously if the scheduling policy was set to "*"
in a profile, e.g.
[scheduler]
group.foo=0:*:1:1:a.out

then Tuned would set the scheduling policy to SCHED_OTHER on RHEL-7,
and to SCHED_RR on Fedora (at least Fedora 27). This is due to the
behaviour of the "chrt" tool. It behaved this way despite what it
says in the following commit, which says the scheduler will not be
changed if set to "*":
https://github.com/redhat-performance/tuned/commit/bf42cb9ba3c

This commit changes/fixes that, so that the scheduling policy is
in fact not changed if "*" is specified.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>