tuned-adm: Fix a traceback when run without arguments
Running tuned-adm without arguments using Python 3 results in
a traceback. This is because in this case parse_args does not
exit with a usage message in Python 3 and the 'action' option
is then undefined. So let's check for this case and print the
usage message ourselves. There may be a better solution to this,
but I wasn't able to come up with anything in a reasonable amount
of time.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>