monitors: Fix adding a hotplugged device to a monitor
If a device, namely a disk, is hotplugged after tuned is started, it
won't be present in Monitor._available_devices. Consequently, it won't
get added to the monitor and so dynamic tuning won't work for the
device. Let's fix that by refreshing the list of available devices when
adding a device to a monitor.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>