sysctl: Set individual parameters by writing to /proc/sys
Set individual sysctl parameters by directly writing to /proc/sys in
order to avoid the overhead of executing the 'sysctl' program for each
setting.

We originally thought about rewriting the functionality using procfs.sysctl
from python-procfs, however I decided to implement it directly, because
it's trivial and because the error handling in procfs.sysctl is not
sufficient for our needs.

The list of deprecated sysctl settings was taken from the sysctl program
from procps-ng.

Fixes #146
Resolves: rhbz#1663412

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>