sysctl: Don't attempt to set a setting if reading the original fails
Previously it could (theoretically) happen that reading the original
value of a setting fails, but then a new value is successfully written.
This would result in Tuned not being able to restore the original value
of the setting.

Even though I don't know of a situation where this could happen in
practice, the code was logically wrong, so let's fix it.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>