cpu: Support specifying fallback CPU scaling governors
The 'governor' option of the 'cpu' plugin now supports specifying
multiple governors. The governors are separated using '|' (the '|'
character is meant to represent a logical 'or' operator; we already use
the same syntax for the 'energy_perf_bias' option). Tuned will set
the first governor that is available on the system.

For example, with the following profile, Tuned will set the 'ondemand'
governor, if it's available. If it's not available, but the 'powersave'
governor is available, 'powersave' will be set. If neither of them are
available, the governor will not be changed.

[cpu]
governor=ondemand|powersave

Resolves: rhbz#1679205

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>