sap-netweaver: Change value of kernel.sem to RHEL-8 default
The default value of kernel.sem in RHEL-8 is '32000 1024000000 500
32000', which is much higher than what we currently have in the profile.
It is unnecessary to lower the settings, the RHEL-8 defaults are safe to
use. So we could simply drop the tuning from the profile, however it
seems better to set the value explicitly so that the profile overrides
any changes that may have been applied to the parameter due to
misconfiguration.

Resolves: rhbz#1701394

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>