sap-netweaver: Use RHEL-8 defaults for kernel.shmall and kernel.shmmax
The kernel.shmall and kernel.shmmax parameters were previously set using
a script that calculated their values based on the system's memory size.
However the script was broken for high values of memory size.

The default value of kernel.shmall and kernel.shmmax in RHEL-8 is
18446744073692774399, which is sufficiently high and it's recommended to
use it. So we could simply drop the tuning from the profile, however it
seems better to set the value explicitly so that the profile overrides
any changes that may have been applied to the parameters due to
misconfiguration.

Resolves: rhbz#1708418

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>